pyproject.toml
[project]
name = "tehlug-tlg-bot"
version = "0.1.0"
description = "Telegram bot for Tehran Linux User Group Q&A system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot[socks]>=20.7",
    "python-dotenv>=1.0.0",
    "httpx[socks]>=0.28.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "asyncpg>=0.31.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
